\name{DrugVsDiseasedata-package}
\alias{DrugVsDiseasedata-package}
\alias{DrugVsDiseasedata}
\docType{package}
\title{
DrugVsDisease Data Package
}
\description{
Data package to provide default ranked disease profiles, clusters and Cytoscape \href{http://www.cytoscape.org/}{http://www.cytoscape.org/} file format for the DrugVsDisease package.
}
\details{
\tabular{ll}{
Package: \tab DrugVsDiseasedata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-16\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
Data loaded automatically in the DrugVsDisease package.
}
\author{
C. Pacini

Maintainer: Julio Saez-Rodriguez{saezrodriguez@ebi.ac.uk}
}
\references{
Disease Profiles:
\href{http://www.ncbi.nlm.nih.gov/geo/}{http://www.ncbi.nlm.nih.gov/geo/}
}

\keyword{ package }
\seealso{

}
\examples{
#load the disease Ranked lists:
data(diseaseRL)
}
